backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.1.0",
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.8.1",
]
